+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.34 - not yet released
 * New features:
   * Exceptions should be shown to the user as a AlertDialog box with copyable text. [requested by Nicolas]
+  * Mapiah should check and present a notice when there is a new version available. [requested by Wal]
 * Fixed bugs:
 * Infrastructure maintenance:
   * flutter upgrade to 3.38.9.